reasoning_attacks/requirements.txt
datasets
einops
nnsight
numpy
pandas
protobuf
PyYAML
scikit-learn
scipy
seaborn
sentencepiece
torch
transformers
tqdm
wandb
matplotlib
huggingface-hub
vllm
git+https://github.com/dsbowen/strong_reject.git@main

